--- a/src/main/java/com/telepathicgrunt/worldblender/utils/CodecCache.java
+++ b/src/main/java/com/telepathicgrunt/worldblender/utils/CodecCache.java
@@ -27,6 +27,7 @@ public class CodecCache<T> {
 
     public void clear() {
         cache.clear();
+        requestCount.set(0);
     }
 
     public Optional<JsonElement> get(T value) {