This PR contains a couple of optimizations around encoding/comparing the configured features:
- Opt#1 - Caching - either due to feature instances being reused in multiple biomes, or WB's own processing, you end up encoding the same feature multiple times which can both eat up time and create a lot of relatively large but short-lived objects. The `CodecCache` adds a simple caching layer around the ConfiguredFeature codec ensuring the encode is only computed once per ConfiguredFeature instance. 
- Opt#2 - Comparing - just some minor reording of the `serializeAndCompareFeature` method can save unecessary computation by front-loading the cheap checks and returning early where possible. I also use `JsonElement#equals` in place of the `toString().equals()` - both result in recursive traversal of the json structure but the latter allocates alsorts to create the string before getting do the equality check, so the former should be noticeably quicker and produces less garbage.

### Testing
I tested using this fork, BYG v1.1.10, and BoP v13.0.0.426

For my baseline (ie pre-changes) 'blend time' I was getting around `13500ms`
With Opt#1 alone it was down to around `5500ms`
With Opt#1+2 active it was running at `3000ms`
Overall speed up of ~10 seconds 👍

The cache computed the json for `594` unique configured feature instances, and had a total of `264,481` requests.
This makes an average saving of over `440` duplicate encodes per configured feature.

### Notes
- Wasn't sure on the codestyle so feel free to beautify as you see fit.
- I've left the time measurement logging in the blendWorlds method. You may want to remove, idk.
- The default cache isn't thread-safe but there is a helper method there to switch to a concurrentmap-backed cache if needed.
- Testing required - I'm not familiar enough with this mod/dimension to be able to spot if I broke something (appears to work though) so this is provided as-is, no guarantees, yadayada XD